refactor(tools): improve new-task tool implementation
This commit refactors the `new-task` tool and related components to improve its functionality and align with the latest changes in the codebase. The key changes include:

- Deleting the legacy `custom-agent.ts` file.
- Introducing `new-task-back.ts` to separate concerns.
- Updating the `new-task.ts` tool with a more robust implementation.
- Modifying related UI components to support the updated tool.
