+++ packages/vscode-webui/src/components/tool-invocation/index.tsx
@@ -89,6 +89,4 @@ const Tools: Record<string, React.FC<ToolProps<any>>> = {
   todoWrite: todoWriteTool,
   // @ts-ignore
   newTask: newTaskTool,
-  // @ts-ignore
-  newCustomAgent: newTaskTool,
 };
