+++ packages/vscode-webui/src/components/tool-invocation/tools/new-task/index.tsx
@@ -15,7 +15,7 @@ import type { ToolProps } from "../../types";
 import { useInlinedSubTask } from "./use-inlined-sub-task";
 import { useLiveSubTask } from "./use-live-sub-task";
 
-interface NewTaskToolProps extends ToolProps<"newTask" | "newCustomAgent"> {
+interface NewTaskToolProps extends ToolProps<"newTask"> {
   // For storybook visualization
   taskThreadSource?: TaskThreadSource;
 }
@@ -41,10 +41,7 @@ export const newTaskTool: React.FC<NewTaskToolProps> = ({
     );
   }
 
-  let agentType: string | undefined;
-  if (tool.type === "tool-newCustomAgent") {
-    agentType = tool.input?.agentType;
-  }
+  const agentType = tool.input?.agentType;
   const toolTitle = agentType ?? "Subtask";
 
   const [showMessageList, setShowMessageList] = useState(false);
