+++ packages/vscode-webui/src/features/settings/components/sections/tools-section.tsx
@@ -47,16 +47,14 @@ const ToolDescriptions: Record<ToolName, string> = {
     "For complex tasks that require multiple steps, Pochi uses the `todoWrite` tool to create and manage a checklist. This helps Pochi organize its work, track its progress, and ensure that it does not miss any steps. It also gives you visibility into its plan and how it's progressing.\n\nPochi will create a list of tasks at the beginning of a complex request and update the status of each task (`pending`, `in-progress`, `completed`) as it works through them. This systematic approach helps Pochi break down large problems into smaller, manageable pieces, and ensures a successful outcome.",
   writeToFile:
     "The `writeToFile` tool is what Pochi uses when it needs to create a new file from scratch or completely replace the contents of an existing one. This is perfect for generating new components, adding new configuration files, or performing large-scale refactors where the majority of a file needs to be changed.\n\nUnlike `applyDiff`, which makes targeted changes, `writeToFile` replaces everything in the file with the new content Pochi provides. It's a powerful tool for making significant additions or modifications to your project. Pochi will always be careful to confirm that overwriting a file is the correct action.",
-  newTask:
-    "When Pochi encounters a task that requires extensive searching or would benefit from dedicated processing, Pochi uses the `newTask` tool. This tool allows Pochi to create a separate, focused task that can run independently while Pochi continues with other work. It's particularly useful for complex code analysis, extensive file searches, or when Pochi needs to perform repetitive operations across many files.\n\nPochi can use `newTask` to delegate specific responsibilities, such as finding all instances of a deprecated function across a large codebase or analyzing project dependencies. This tool enhances Pochi's efficiency by allowing parallel processing of multiple concerns.",
   startBackgroundJob:
     "The `startBackgroundJob` tool allows Pochi to initiate long-running processes that don't need to block the main task flow. This is ideal for starting development servers, file watchers, continuous build processes, or other persistent tasks that should continue running in the background.\n\nFor example, Pochi can use this tool to start a development server while simultaneously working on frontend components. The background job continues to run, and Pochi can later check its status or output using other tools. This enables Pochi to manage complex development environments with multiple concurrent processes.",
   readBackgroundJobOutput:
     "After starting a background job, Pochi uses the `readBackgroundJobOutput` tool to monitor its progress and retrieve any output or logs it has generated. This tool allows Pochi to check the status of background processes, read error messages, view progress updates, or capture the results of long-running operations.\n\nPochi can use this tool periodically to ensure background processes are running correctly, or to gather information needed for subsequent steps. It provides real-time visibility into asynchronous operations, helping Pochi make informed decisions based on the background job's current state.",
   killBackgroundJob:
     "When a background process is no longer needed or is causing issues, Pochi uses the `killBackgroundJob` tool to terminate it cleanly. This is important for resource management and ensuring that processes don't continue running unnecessarily after their purpose has been fulfilled.\n\nPochi might use this tool to stop development servers after completing work, cancel long-running builds, or terminate processes that are consuming too many resources. It gives Pochi full control over the lifecycle of background processes, ensuring efficient and clean task completion.",
-  newCustomAgent:
-    "The `newCustomAgent` tool allows Pochi to create a new custom agent instance. This is useful for tasks that require a dedicated agent with specific capabilities or configurations. By using this tool, Pochi can tailor the agent's behavior to better suit the needs of the task at hand.",
+  newTask:
+    "The `newTask` tool allows Pochi to create a new task with a dedicated agent. This is useful for tasks that require a dedicated agent with specific capabilities or configurations. By using this tool, Pochi can tailor the agent's behavior to better suit the needs of the task at hand.",
 };
 
 const AllTools = keys({ ...createClientTools() })
